requirements.txt
atom
enaml
h5py>=2.9.0
lmfit
matplotlib
numpy>=1.15
pandas
pillow
pystackreg
pyyaml
requests
scikit-image
scipy
six
tifffile
